# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: CRM Platform API
  description: |
    Comprehensive API documentation for the CRM Platform microservices including
    authentication, user management, and security features.
    
    ## Authentication
    
    All protected endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Multi-Tenancy
    
    Most endpoints require a tenant context header:
    ```
    X-Tenant-ID: <tenant_uuid>
    ```
    
    ## Rate Limiting
    
    API endpoints are rate limited. Check response headers for current limits:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Window reset time (Unix timestamp)
    
  version: 1.0.0
  contact:
    name: CRM Platform API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Authentication Service (Development)
  - url: http://localhost:8082/api/v1
    description: User Management Service (Development)
  - url: http://localhost:8083/api/v1
    description: Tenant Service (Development)
  - url: https://api.crm-platform.com/v1
    description: Production Environment
security:
  - BearerAuth: []
  - TenantContext: []
paths:
...
  /roles:
    get:
      tags:
        - Role Management
      summary: List Roles
      description: Retrieve all available roles
      operationId: listRoles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'

    post:
      tags:
        - Role Management
      summary: Create Role
      description: Create a new role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
    TenantContext:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant context identifier
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (0-based)
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    CreateRoleRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: "SALES_MANAGER"
        description:
          type: string
          example: "Sales team manager with CRM access"
        permissions:
          type: array
          items:
            type: string
    RoleResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Role'
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
        user_count:
          type: integer
        created_at:
          type: string
          format: date-time
        is_system_role:
          type: boolean
    RoleListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PagedResponse'

    # Team Management Schemas
...
    PagedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        total_elements:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
...